+++ data/src/main/java/com/ssing/data/lesson/remote/dto/response/InstructorLessonDetailCanceledResponse.kt
@@ -20,7 +20,7 @@ internal data class InstructorLessonDetailCanceledResponse(
 
     @Serializable
     internal data class CanceledBy(
-        @SerialName("memberId") val memberId: Int,
+        @SerialName("memberId") val memberId: Long,
         @SerialName("name") val name: String,
     )
 
@@ -43,8 +43,8 @@ internal data class InstructorLessonDetailCanceledResponse(
 
     @Serializable
     internal data class MatchingRequest(
-        @SerialName("matchingRequestId") val matchingRequestId: Int,
-        @SerialName("representativeMemberId") val representativeMemberId: Int,
+        @SerialName("matchingRequestId") val matchingRequestId: Long,
+        @SerialName("representativeMemberId") val representativeMemberId: Long,
         @SerialName("representativeMemberName") val representativeMemberName: String,
         @SerialName("headcount") val headcount: Int,
         @SerialName("teamLessonPrice") val teamLessonPrice: Int,
@@ -53,7 +53,7 @@ internal data class InstructorLessonDetailCanceledResponse(
 
     @Serializable
     internal data class Participant(
-        @SerialName("participantId") val participantId: Int,
+        @SerialName("participantId") val participantId: Long,
         @SerialName("gender") val gender: String,
         @SerialName("age") val age: Int,
     )
